Pull TestSegmentCacheManager into its own class for reusability.
