+++ server/src/test/java/org/apache/druid/server/coordination/SegmentBootstrapperTest.java
@@ -20,26 +20,15 @@
 package org.apache.druid.server.coordination;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
-import org.apache.druid.client.coordinator.CoordinatorClient;
-import org.apache.druid.client.coordinator.NoopCoordinatorClient;
 import org.apache.druid.guice.ServerTypeConfig;
 import org.apache.druid.java.util.common.Intervals;
-import org.apache.druid.java.util.common.MapUtils;
-import org.apache.druid.java.util.common.concurrent.Execs;
-import org.apache.druid.java.util.common.concurrent.ScheduledExecutorFactory;
 import org.apache.druid.java.util.emitter.EmittingLogger;
 import org.apache.druid.java.util.metrics.StubServiceEmitter;
-import org.apache.druid.segment.ReferenceCountingSegment;
-import org.apache.druid.segment.SegmentLazyLoadFailCallback;
-import org.apache.druid.segment.loading.NoopSegmentCacheManager;
 import org.apache.druid.segment.loading.SegmentLoaderConfig;
 import org.apache.druid.segment.loading.StorageLocationConfig;
-import org.apache.druid.segment.loading.TombstoneSegmentizerFactory;
 import org.apache.druid.server.SegmentManager;
 import org.apache.druid.server.TestSegmentUtils;
 import org.apache.druid.timeline.DataSegment;
-import org.joda.time.Interval;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
@@ -49,15 +38,10 @@ import org.junit.rules.TemporaryFolder;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.ScheduledThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
 
 public class SegmentBootstrapperTest
 {
@@ -66,7 +50,6 @@ public class SegmentBootstrapperTest
   private TestDataSegmentAnnouncer segmentAnnouncer;
   private TestDataServerAnnouncer serverAnnouncer;
   private SegmentLoaderConfig segmentLoaderConfig;
-  private ScheduledExecutorFactory scheduledExecutorFactory;
   private TestCoordinatorClient coordinatorClient;
   private StubServiceEmitter serviceEmitter;
 
@@ -118,19 +101,6 @@ public class SegmentBootstrapperTest
       }
     };
 
-    scheduledExecutorFactory = (corePoolSize, nameFormat) -> {
-      // Override normal behavior by adding the runnable to a list so that you can make sure
-      // all the shceduled runnables are executed by explicitly calling run() on each item in the list
-      return new ScheduledThreadPoolExecutor(corePoolSize, Execs.makeThreadFactory(nameFormat))
-      {
-        @Override
-        public ScheduledFuture<?> schedule(Runnable command, long delay, TimeUnit unit)
-        {
-          return null;
-        }
-      };
-    };
-
     coordinatorClient = new TestCoordinatorClient();
     serviceEmitter = new StubServiceEmitter();
     EmittingLogger.registerEmitter(serviceEmitter);
@@ -151,7 +121,11 @@ public class SegmentBootstrapperTest
 
     final TestSegmentCacheManager cacheManager = new TestSegmentCacheManager(segments);
     final SegmentManager segmentManager = new SegmentManager(cacheManager);
-    final SegmentLoadDropHandler handler = initSegmentLoadDropHandler(segmentManager);
+    final SegmentLoadDropHandler handler = new SegmentLoadDropHandler(
+        segmentLoaderConfig,
+        segmentAnnouncer,
+        segmentManager
+    );
     final SegmentBootstrapper bootstrapper = new SegmentBootstrapper(
         handler,
         segmentLoaderConfig,
@@ -188,9 +162,10 @@ public class SegmentBootstrapperTest
     Assert.assertEquals(1, cacheManager.observedShutdownBootstrapCount.get());
   }
 
+  @Test
   public void testLoadCache() throws Exception
   {
-    Set<DataSegment> segments = new HashSet<>();
+    final Set<DataSegment> segments = new HashSet<>();
     for (int i = 0; i < COUNT; ++i) {
       segments.add(TestSegmentUtils.makeSegment("test" + i, "1", Intervals.of("P1d/2011-04-01")));
       segments.add(TestSegmentUtils.makeSegment("test" + i, "1", Intervals.of("P1d/2011-04-02")));
@@ -209,7 +184,11 @@ public class SegmentBootstrapperTest
 
     final TestSegmentCacheManager cacheManager = new TestSegmentCacheManager(segments);
     final SegmentManager segmentManager = new SegmentManager(cacheManager);
-    final SegmentLoadDropHandler handler = initSegmentLoadDropHandler(segmentManager);
+    final SegmentLoadDropHandler handler = new SegmentLoadDropHandler(
+        segmentLoaderConfig,
+        segmentAnnouncer,
+        segmentManager
+    );
     final SegmentBootstrapper bootstrapper = new SegmentBootstrapper(
         handler,
         segmentLoaderConfig,
@@ -261,7 +240,11 @@ public class SegmentBootstrapperTest
     final TestCoordinatorClient coordinatorClient = new TestCoordinatorClient(segments);
     final TestSegmentCacheManager cacheManager = new TestSegmentCacheManager();
     final SegmentManager segmentManager = new SegmentManager(cacheManager);
-    final SegmentLoadDropHandler handler = initSegmentLoadDropHandler(segmentManager, coordinatorClient);
+    final SegmentLoadDropHandler handler = new SegmentLoadDropHandler(
+        segmentLoaderConfig,
+        segmentAnnouncer,
+        segmentManager
+    );
     final SegmentBootstrapper bootstrapper = new SegmentBootstrapper(
         handler,
         segmentLoaderConfig,
@@ -302,8 +285,11 @@ public class SegmentBootstrapperTest
   {
     final TestSegmentCacheManager cacheManager = new TestSegmentCacheManager();
     final SegmentManager segmentManager = new SegmentManager(cacheManager);
-
-    final SegmentLoadDropHandler handler = initSegmentLoadDropHandler(segmentManager, new NoopCoordinatorClient());
+    final SegmentLoadDropHandler handler = new SegmentLoadDropHandler(
+        segmentLoaderConfig,
+        segmentAnnouncer,
+        segmentManager
+    );
     final SegmentBootstrapper bootstrapper = new SegmentBootstrapper(
         handler,
         segmentLoaderConfig,
@@ -330,141 +316,4 @@ public class SegmentBootstrapperTest
 
     bootstrapper.stop();
   }
-
-  private SegmentLoadDropHandler initSegmentLoadDropHandler(
-      SegmentManager segmentManager,
-      CoordinatorClient coordinatorClient
-  )
-  {
-    return initSegmentLoadDropHandler(segmentLoaderConfig, segmentManager, coordinatorClient);
-  }
-
-  private SegmentLoadDropHandler initSegmentLoadDropHandler(SegmentManager segmentManager)
-  {
-    return initSegmentLoadDropHandler(segmentLoaderConfig, segmentManager, coordinatorClient);
-  }
-
-  private SegmentLoadDropHandler initSegmentLoadDropHandler(
-      SegmentLoaderConfig config,
-      SegmentManager segmentManager,
-      CoordinatorClient coordinatorClient
-  )
-  {
-    return new SegmentLoadDropHandler(
-        config,
-        segmentAnnouncer,
-        segmentManager,
-        scheduledExecutorFactory.create(5, "SegmentLoadDropHandlerTest-[%d]")
-    );
-  }
-
-  /**
-   * A local cache manager to test the bootstrapping and segment add/remove operations. It stubs only the necessary
-   * methods to support these operations; any other method invoked will throw an exception from the base class,
-   * {@link NoopSegmentCacheManager}.
-   */
-  private static class TestSegmentCacheManager extends NoopSegmentCacheManager
-  {
-    private final List<DataSegment> cachedSegments;
-
-    private final List<DataSegment> observedBootstrapSegments;
-    private final List<DataSegment> observedBootstrapSegmentsLoadedIntoPageCache;
-    private final List<DataSegment> observedSegments;
-    private final List<DataSegment> observedSegmentsLoadedIntoPageCache;
-    private final List<DataSegment> observedSegmentsRemovedFromCache;
-    private final AtomicInteger observedShutdownBootstrapCount;
-
-    TestSegmentCacheManager()
-    {
-      this(ImmutableSet.of());
-    }
-
-    TestSegmentCacheManager(final Set<DataSegment> segmentsToCache)
-    {
-      this.cachedSegments = ImmutableList.copyOf(segmentsToCache);
-      this.observedBootstrapSegments = new ArrayList<>();
-      this.observedBootstrapSegmentsLoadedIntoPageCache = new ArrayList<>();
-      this.observedSegments = new ArrayList<>();
-      this.observedSegmentsLoadedIntoPageCache = new ArrayList<>();
-      this.observedSegmentsRemovedFromCache = new ArrayList<>();
-      this.observedShutdownBootstrapCount = new AtomicInteger(0);
-    }
-
-    @Override
-    public boolean canHandleSegments()
-    {
-      return true;
-    }
-
-    @Override
-    public List<DataSegment> getCachedSegments()
-    {
-      return cachedSegments;
-    }
-
-    @Override
-    public ReferenceCountingSegment getBootstrapSegment(DataSegment segment, SegmentLazyLoadFailCallback loadFailed)
-    {
-      observedBootstrapSegments.add(segment);
-      return getSegmentInternal(segment);
-    }
-
-    @Override
-    public ReferenceCountingSegment getSegment(final DataSegment segment)
-    {
-      observedSegments.add(segment);
-      return getSegmentInternal(segment);
-    }
-
-    private ReferenceCountingSegment getSegmentInternal(final DataSegment segment)
-    {
-      if (segment.isTombstone()) {
-        return ReferenceCountingSegment
-            .wrapSegment(TombstoneSegmentizerFactory.segmentForTombstone(segment), segment.getShardSpec());
-      } else {
-        return ReferenceCountingSegment.wrapSegment(
-            new TestSegmentUtils.SegmentForTesting(
-                segment.getDataSource(),
-                (Interval) segment.getLoadSpec().get("interval"),
-                MapUtils.getString(segment.getLoadSpec(), "version")
-            ), segment.getShardSpec()
-        );
-      }
-    }
-
-    @Override
-    public void loadSegmentIntoPageCache(DataSegment segment)
-    {
-      observedSegmentsLoadedIntoPageCache.add(segment);
-    }
-
-    @Override
-    public void loadSegmentIntoPageCacheOnBootstrap(DataSegment segment)
-    {
-      observedBootstrapSegmentsLoadedIntoPageCache.add(segment);
-    }
-
-    @Override
-    public void shutdownBootstrap()
-    {
-      observedShutdownBootstrapCount.incrementAndGet();
-    }
-
-    @Override
-    public void storeInfoFile(DataSegment segment)
-    {
-    }
-
-    @Override
-    public void removeInfoFile(DataSegment segment)
-    {
-    }
-
-    @Override
-    public void cleanup(DataSegment segment)
-    {
-      observedSegmentsRemovedFromCache.add(segment);
-    }
-  }
-
 }
